Increase NVRAM access timeout to five minutes
Apparently, three minutes is not completely unexpected... use five minutes to
give some padding.  It's possible we'll have to revisit this since only one
HWRM command can be issued at a time... maybe a HWRM lock timeout?

+++ sys/dev/bnxt/bnxt.h
@@ -497,7 +497,7 @@ struct bnxt_softc {
 	struct bnxt_vf_info	vf;
 
 	uint16_t		hwrm_cmd_seq;
-	uint16_t		hwrm_cmd_timeo;	/* milliseconds */
+	uint32_t		hwrm_cmd_timeo;	/* milliseconds */
 	struct iflib_dma_info	hwrm_cmd_resp;
 	/* Interrupt info for HWRM */
 	struct if_irq		irq;
